Move contents of Home to List
