+++ src/views/List.jsx
@@ -1,5 +1,55 @@
 import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { SingleList } from '../components';
+import { createList } from '../api/firebase';
 
-export function List({ data, listPath }) {
-	return <></>;
+export function List({ user, data, setListPath }) {
+	const [shoppingListName, setShoppingListName] = useState('');
+	const navigate = useNavigate(); // useNavigate doc suggests using redirect
+
+	const handleSubmit = async (e) => {
+		e.preventDefault();
+		try {
+			const newListPath = await createList(
+				user.uid,
+				user.email,
+				shoppingListName,
+			);
+			setListPath(newListPath);
+			alert('You have created a new shopping list: ' + shoppingListName);
+			navigate('/list');
+		} catch (error) {
+			alert(`Your list was not created, please try again. Error: ${error}`);
+		}
+	};
+	return (
+		<>
+			<ul>
+				{data && data.length > 0 ? (
+					data.map((list) => (
+						<SingleList
+							key={list.name}
+							name={list.name}
+							setListPath={setListPath}
+							path={list.path}
+						/>
+					))
+				) : (
+					<h1>You have no Lists!</h1>
+				)}
+			</ul>
+			<form onSubmit={handleSubmit}>
+				<label>
+					Shopping List:
+					<input
+						type="text"
+						value={shoppingListName}
+						id="shopping-list"
+						onChange={(e) => setShoppingListName(e.target.value)}
+					></input>
+				</label>
+				<input type="submit" />
+			</form>
+		</>
+	);
 }
